% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataApis.R
\name{api_all}
\alias{api_all}
\title{API all}
\usage{
api_all(df_serie, from = "2000-01-01", to = "2024-11-30", same.length = F)
}
\arguments{
\item{df_serie}{Un dataframe con las series a cargar}

\item{from}{Fecha inicial que nos interesa escrita en string, ej. "2022-01-01"}

\item{to}{Fecha final que nos interesa escrita en string}

\item{same.length}{Está puesto en False por defecto; sin embargo, si tus series son de la misma temporalidad
y longitud, puedes ponerlo en True para obtener un dataframe combinado con columnas nombradas según "description".}
}
\value{
Si same.length=F regresará una lista con los dataframes de cada serie,
si es T, entonces regresará el dataframe combinado de todas las series (necesita misma longitud).
}
\description{
Obtén múltiples series al mismo tiempo desde un df.
}
\details{
\itemize{
\item Recuerda primero introducir los tokens necesarios con set_api_tokens()
\item Necesitas tener un dataframe con las siguientes columnas en el siguiente orden:
\itemize{
\item "id" (id de la serie)
\item "description" (nombre a usar para la serie),
\item "origen" (institución de la que viene) el origen tiene que estar escrito en mayusculas, ejemplo "BANXICO" o "BIE"
\item "temp" (este es para la API de INEGI que necesita estipular mes o trimestre) y
\item opcionalmente la columna "do" que aplica transformaciones con do_transform().
}
}

Descripción más detallada para cargar varias series disponibles en el github del paquete: https://github.com/rojoaldrette/DataAPIs
}
\examples{
series <- read.csv("modelo_1.csv")
var_x <- api_one.banxico(series, from="2000-01-01", to="2020-01-01")
}
