% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataApis.R
\name{api_one.banxico}
\alias{api_one.banxico}
\title{API Banxico}
\usage{
api_one.banxico(id, from = "2000-01-01", to = "2025-06-01")
}
\arguments{
\item{id}{El id de la serie de Banxico}

\item{from}{Fecha inicial que nos interesa escrita en string, ej. "2022-01-01"}

\item{to}{Fecha final que nos interesa escrita en string}
}
\value{
Un DataFrame con dos columnas: "fecha" y "valor"
}
\description{
Obtén serie de la API de banxico
}
\examples{
inflacion <- api_one.banxico("SP1", from="2000-01-01", to="2020-01-01");
DeudaPub <- api_one.banxico("SR2632");
}
