% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataApis.R
\name{api_one.fred}
\alias{api_one.fred}
\title{API FRED}
\usage{
api_one.fred(id, from = "2000-01-01", to = "2024-11-30")
}
\arguments{
\item{id}{El id de la serie de Banxico}

\item{from}{Fecha inicial que nos interesa escrita en string, ej. "2022-01-01"}

\item{to}{Fecha final que nos interesa escrita en string}
}
\value{
Un DataFrame con dos columnas: "fecha" y "valor"
}
\description{
Obtén series de la API de la Federal Reserve Economic Data
}
\details{
*Recuerda primero introducir tu token de la FRED con set_api_tokens()
}
\examples{
us_inpc <- api_one.fred("CPIAUCNS", from="2000-01-01", to="2020-01-01")
}
