% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataApis.R
\name{api_one.inegi_bise}
\alias{api_one.inegi_bise}
\title{API INEGI BISE}
\usage{
api_one.inegi_bise(
  id,
  temporalidad = "m",
  from = "2000-01-01",
  to = "2024-11-01"
)
}
\arguments{
\item{id}{El id de la serie del BIE}

\item{temporalidad}{Acepta "m" (month) o "q" (quarter) dependiendo de la serie,
por defecto está en "m", ya que es la más común en el BIE.}

\item{from}{Fecha inicial que nos interesa escrita en string, ej. "2022-01-01"}

\item{to}{Fecha final que nos interesa escrita en string}
}
\value{
Un DataFrame con dos columnas: "fecha" y "valor"
}
\description{
Obtén series de la API de INEGI del Banco de Indicadores Socio-Económicos
}
\details{
*Recuerda primero introducir tu token de INEGI con set_api_tokens()
}
\examples{
consumo <- api_one.inegi_bise("740933", from="2000-01-01", to="2020-01-01",
                              temporalidad="m");
}
