% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataApis.R
\name{api_one.yahoo}
\alias{api_one.yahoo}
\title{API INEGI BIE}
\usage{
api_one.yahoo(id, from = "2000-01-01", to = "2024-11-30", adjusted = T)
}
\arguments{
\item{id}{El id de la serie de yahoo}

\item{from}{Fecha inicial que nos interesa escrita en string, ej. "2022-01-01"}

\item{to}{Fecha final que nos interesa escrita en string}
}
\value{
Un DataFrame con dos columnas: "fecha" y "valor"
}
\description{
Obtén series de la API de Yahoo Finanzas
}
\details{
\itemize{
\item No necesita token, lo hace a través del paquete quantmod
\item Si necesitas mayor precisión (apertura, cierre, etc..) recomiendo usar directamente
quantmod, ofrece mayor personalización y accesibilidad a datos financieros
}
}
\examples{
consumo <- api_one.yahoo("MXN=X", from="2005-01-01", to="2008-01-01")
}
