% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{set_api_tokens}
\alias{set_api_tokens}
\title{Set API tokens for the package}
\usage{
set_api_tokens(banxico_token = NULL, inegi_token = NULL, fred_token = NULL)
}
\arguments{
\item{banxico_token}{Banxico API token (character)}

\item{inegi_token}{INEGI API token (character)}

\item{fred_token}{FRED API token (character)}
}
\value{
Invisible NULL
}
\description{
Set API tokens for the package
}
\examples{
set_api_tokens(
  banxico_token = "your_banxico_token",
  inegi_token = "your_inegi_token",
  fred_token = "your_fred_token"
)
}
